going to just implement the efficient chunk reader
it will be a new function on the file logger that returns an iterator
function. the function returns an io.Reader or an error. The goal is to
automatically leverage sendfile when reading a static file. If we can
return file partitions, the reader will be an *os.File at an arbitrary
seek position. it may also be wrapped in an io.LimitReader. Otherwise
it'll be a bufio.Reader backed by the head partition.

protocol-wise, it is wrapped in the chunk envelope:

+<bytes>\r\n
...
+<bytes>\r\n
...
+EOF\r\n

OH OH the last reader should close all the file handles in the iterator
and return io.EOF.