use query index for reads
prevents hitting disk twice in an operation

preallocate and batch allocs

wip, fixed the issue

it was that the server was returning 1 batch when there were actually
multiple batches. this commit fixes that by having it return the correct
number of batches.

now i want to make sure the server is not sending the rest of the unsent
batch on subsequent requests

do a binary search lookup instead of the naive way

unit test query index

new index wired up, persisting index to disk on rotate

lookups are still not correct though

use parts map to save iterations